requirements.txt
numpy==2.1.2
Pillow==10.4.0
opencv-python==4.10.0.84
cryptography==43.0.0
click==8.1.7
Flask==3.0.2